+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Audit log (per-project) with possibility to browse, filter, and clear using 
   a simple dialog in client-side
 
+### Changed
+
+- Prepared getting and passing SHACL specs from FDP to OpenRefine frontend
+
 ## [1.2.0]
 
 ### Added